@@ -1,10 +1,11 @@
-﻿#if NEMESIS_BINARY_PACKAGE_TESTS
-using System.Diagnostics;
+﻿using System.Diagnostics;
+
+#if NEMESIS_BINARY_PACKAGE_TESTS
 using Nemesis.Essentials.Runtime;
 
 namespace Nemesis.Essentials.Tests.Runtime;
 #else
-namespace Nemesis.Essentials.Sources.Tests.Runtime
+namespace Nemesis.Essentials.Sources.Tests.Runtime;
 #endif
 
 [TestFixture]
