@@ -2,13 +2,12 @@
 using System.Reflection;
 using System.Runtime.CompilerServices;
 
-
 #if NEMESIS_BINARY_PACKAGE_TESTS
 using Nemesis.Essentials.Runtime;
 
 namespace Nemesis.Essentials.Tests.Runtime;
 #else
-namespace Nemesis.Essentials.Sources.Tests.Runtime
+namespace Nemesis.Essentials.Sources.Tests.Runtime;
 #endif
 
 [TestFixture]
